requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
django-debug-toolbar==5.1.0
django-extensions==4.1
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.0.2
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.2.0
mysqlclient==2.2.7
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
PyMySQL==1.1.1
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
yarl==1.18.3
